bugfix: null value
